added clear chats,unread,etc features

+++ src/features/chats/sections/chatList/components/ui/popover.tsx
@@ -1,22 +1,27 @@
-import ButtonItem from "../../../components/buttonItem";
+import ButtonItem from "../../../../components/buttonItem";
 import { Popover, Typography } from "@mui/material";
 import MarkEmailUnreadIcon from "@mui/icons-material/MarkEmailUnread";
 import FavoriteBorderIcon from "@mui/icons-material/FavoriteBorder";
 import FavoriteIcon from "@mui/icons-material/Favorite";
 import DeleteIcon from "@mui/icons-material/Delete";
 import BlockTwoToneIcon from "@mui/icons-material/BlockTwoTone";
-import { ChatListTypes } from "../../../types/chatTypes";
+import { ChatListTypes } from "../../../../types/chatTypes";
 import { Dispatch, SetStateAction, useEffect, useState } from "react";
-import { useAuth } from "../../../../auth/context/authContext";
-import { deleteChat, toggleFavouriteChat } from "../../../services/chatService";
-import AlertMessage from "../../../../../shared/components/layout/alertMessage";
-import { useSelectedUserContext } from "../../../context/selectedUserContext";
-import ConfirmationModal from "../../../components/layout/confirmationModal";
+import { useAuth } from "../../../../../auth/context/authContext";
+import {
+  clearChat,
+  deleteChat,
+  toggleFavouriteChat,
+} from "../../../../services/chatService";
+import AlertMessage from "../../../../../../shared/components/layout/alertMessage";
+import RemoveCircleOutlineOutlinedIcon from "@mui/icons-material/RemoveCircleOutlineOutlined";
+import { useSelectedUserContext } from "../../../../context/selectedUserContext";
+import ConfirmationModal from "../../../../components/layout/confirmationModal";
 import {
   blockUser,
   unBlockUser,
-} from "../../../../../shared/services/firebase/userService";
-import useActiveUser from "../../../../../shared/hooks/useActiveUser";
+} from "../../../../../../shared/services/firebase/userService";
+import useActiveUser from "../../../../../../shared/hooks/useActiveUser";
 
 type propsTypes = {
   openPopover: boolean;
@@ -38,17 +43,17 @@ const PopOver = ({
   const [showAlert, setShowAlert] = useState<boolean>(false);
   const [alertMessage, setAlertMessage] = useState<string>("");
   const { setSelectedChat } = useSelectedUserContext();
-  const [modalAction, setModalAction] = useState<"block" | "delete" | null>(
-    null
-  );
-  const { activeUser } = useActiveUser();
-  const [isBlocked, setIsBlocked] = useState(false);
-
-  useEffect(() => {
-    if (activeUser && chat) {
-      setIsBlocked(activeUser.blockedUsers.includes(chat.otherUid));
-    }
-  }, [activeUser, chat]);
+  const [modalAction, setModalAction] = useState<
+    "block" | "delete" | "clear" | null
+  >(null);
+  const { isBlocked } = useActiveUser(chat?.otherUid!);
+  const modalTextMap: Record<string, string> = {
+    block: `Are you sure you want to ${
+      isBlocked ? "unblock" : "block"
+    } this user?`,
+    delete: "Are you sure you want to delete this chat?",
+    clear: "Are you sure you want to clear this chat?",
+  };
 
   useEffect(() => {
     console.log(chat);
@@ -77,32 +82,33 @@ const PopOver = ({
     if (!user?.uid || !chat?.id) return;
     if (isBlocked) {
       unBlockUser(user?.uid, chat.otherUid);
-      setIsBlocked(false);
+      setAlertMessage("user is unBlocked");
     } else {
       blockUser(user?.uid, chat.otherUid);
-      setIsBlocked(true);
+      setAlertMessage("user is blocked");
     }
+    setShowAlert(true);
+  };
+
+  const handleClearChat = async () => {
+    if (!chat?.id) return;
+    setSelectedChat(null);
+    await clearChat(chat.id);
+    setAlertMessage("Successfully Cleared the Chat!");
+    setShowAlert(true);
   };
 
   return (
     <>
       <ConfirmationModal
-        text={
-          modalAction === "block"
-            ? "Are u sure u want to block this user?"
-            : "Are u sure u want to delete chat"
-        }
+        text={modalAction ? modalTextMap[modalAction] : ""}
         open={!!modalAction}
         setOpen={() => setModalAction(null)}
         onConfirm={() => {
-          if (modalAction === "delete") {
-            handleDeleteChat();
-            setModalAction(null);
-          }
-          if (modalAction === "block") {
-            handleBlockUser();
-            setModalAction(null);
-          }
+          if (modalAction === "delete") handleDeleteChat();
+          if (modalAction === "block") handleBlockUser();
+          if (modalAction === "clear") handleClearChat();
+          setModalAction(null);
         }}
       />
       <AlertMessage
@@ -125,10 +131,14 @@ const PopOver = ({
         }}
       >
         <ButtonItem
-          icon={<MarkEmailUnreadIcon />}
-          text="Mark as unread"
+          icon={<RemoveCircleOutlineOutlinedIcon />}
+          text="Clear chat"
           isMenu={true}
           className="flex items-center gap-4 w-full p-2 hover:bg-[#e4e2e2]"
+          handleClick={() => {
+            setModalAction("clear");
+            handleClose();
+          }}
         />
         <ButtonItem
           icon={isFavourite ? <FavoriteIcon /> : <FavoriteBorderIcon />}
